Frequency sweep
I'm writing this commit log a couple of weeks after I wrote the code,
so I don't quite recall what this one did; I just had noted that it's
a frequency sweep.  IIRC, this one was doing an FM transmission with
the smallest possible deviation (i.e., only toggling the lowest bit of
OSCTUNE) and changed the speed of the delay loop to sweep the AF
frequency.